ai-trading-bot/requirements.txt
yfinance>=0.2.30
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
flask>=3.0
pyyaml>=6.0
requests>=2.31
schedule>=1.2
rich>=13.0
python-dotenv>=1.0
joblib>=1.3
